Remove old methods from rc-zip-sync
