+++ rc-zip-sync/Cargo.toml
@@ -21,26 +21,18 @@ path = "examples/jean.rs"
 
 [dependencies]
 positioned-io = { version = "0.3.3", optional = true }
-flate2 = { version = "1.0.28", optional = true }
 rc-zip = { version = "2.0.1", path = "../rc-zip" }
-lzma-rs = { version = "0.3.0", features = ["stream"], optional = true }
-deflate64 = { version = "0.1.7", optional = true }
-bzip2 = { version = "0.4.4", optional = true }
-zstd = { version = "0.13.0", optional = true }
 oval = "2.0.0"
-crc32fast = "1.3.2"
 tracing = "0.1.40"
-cfg-if = "1.0.0"
-winnow = "0.5.36"
 
 [features]
 default = ["file", "deflate"]
 file = ["positioned-io"]
-deflate = ["dep:flate2", "rc-zip/deflate"]
-deflate64 = ["dep:deflate64"]
-lzma = ["dep:lzma-rs"]
-bzip2 = ["dep:bzip2"]
-zstd = ["dep:zstd"]
+deflate = ["rc-zip/deflate"]
+deflate64 = ["rc-zip/deflate64"]
+lzma = ["rc-zip/lzma"]
+bzip2 = ["rc-zip/bzip2"]
+zstd = ["rc-zip/zstd"]
 
 [dev-dependencies]
 chrono = "0.4.33"
@@ -50,3 +42,4 @@ indicatif = "0.17.7"
 test-log = { version = "0.2.14", default-features = false, features = ["tracing-subscriber", "trace"] }
 tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
 rc-zip = { version = "2.0.1", path = "../rc-zip", features = ["corpus"] }
+cfg-if = "1.0.0"
