+++ rc-zip-sync/src/lib.rs
@@ -7,17 +7,6 @@
 
 #![warn(missing_docs)]
 
-macro_rules! transition {
-    ($state: expr => ($pattern: pat) $body: expr) => {
-        $state = if let $pattern = std::mem::take(&mut $state) {
-            $body
-        } else {
-            unreachable!()
-        };
-    };
-}
-
-mod decoder;
 mod entry_reader;
 mod read_zip;
