+++ rc-zip-tokio/Cargo.toml
@@ -22,14 +22,15 @@ tokio = { version = "1.35.1", features = ["fs", "io-util", "rt-multi-thread"] }
 futures = { version = "0.3.30" }
 pin-project-lite = { version = "0.2.13" }
 oval = "2.0.0"
-crc32fast = "1.3.2"
 tracing = "0.1.40"
-cfg-if = "1.0.0"
-winnow = "0.5.36"
 
 [features]
 default = ["deflate"]
 deflate = ["rc-zip/deflate"]
+deflate64 = ["rc-zip/deflate64"]
+lzma = ["rc-zip/lzma"]
+bzip2 = ["rc-zip/bzip2"]
+zstd = ["rc-zip/zstd"]
 
 [dev-dependencies]
 test-log = { version = "0.2.14", default-features = false, features = ["tracing-subscriber", "trace"] }
